+++ example/app.json
@@ -1,16 +1,15 @@
 {
   "expo": {
-    "name": "expo-qr-notify-example",
+    "name": "react-native-store-version-example",
     "description": "This project is really great.",
-    "slug": "expo-qr-notify-example",
+    "slug": "react-native-store-version",
     "privacy": "public",
-    "sdkVersion": "31.0.0",
+    "sdkVersion": "37.0.0",
     "platforms": [
       "ios",
       "android"
     ],
-    "primaryColor": "#ff0000",
-    "version": "1.0.0",
+    "version": "1.3.0",
     "orientation": "portrait",
     "icon": "./assets/icon.png",
     "splash": {
@@ -25,23 +24,7 @@
       "**/*"
     ],
     "ios": {
-      "supportsTablet": true,
-      "bundleIdentifier": "enter_your_bundleIdentifier"
-    },
-    "android": {
-      "package": "enter_your_package_name"
-    },
-    "hooks": {
-      "postPublish": [
-        {
-          "file": "expo-qr-notify",
-          "config": {
-            "slack_webhook": "enter_your_slack_incoming_webhook_url",
-            "slack_channel": "#general",
-            "git_service_token": "enter_your_github_token"
-          }
-        }
-      ]
+      "supportsTablet": true
     }
   }
 }
